app.py
import os

from flask import Flask
from flask_smorest import Api
from flask_migrate import Migrate
from dotenv import load_dotenv

# resources holds blueprint / api routes and functions
from db import db
import models

from resources.letter import blp as LetterBluePrint
from resources.alphabet import blp as AlphabetBluePrint

def create_app(db_url=None):
    app = Flask(__name__)
    load_dotenv()

    app.config["PROPAGATE_EXCEPTONS"] = True
    app.config["API_TITLE"] = "CMW REST API"
    app.config["API_VERSION"] = "v1"
    app.config["OPENAPI_VERSION"] = "3.0.3"
    app.config["OPENAPI_URL_PREFIX"] = "/"
    app.config["OPENAPI_SWAGGER_UI_PATH"] = "/swagger-ui"
    app.config["OPENAPI_SWAGGER_UI_URL"] = "https://cdn.jsdelivr.net/npm/swagger-ui-dist/"
    app.config["SQLALCHEMY_DATABASE_URI"] = db_url or os.getenv("DATABASE_URL", "sqlite:///data.db") # used for development
    app.config["SQLALCHEMY_TRACK_MODIFICATIONS"] = False

    db.init_app(app)
    with app.app_context():
        db.create_all()
    # migrate = Migrate(app, db)
    api = Api(app)

    api.register_blueprint(LetterBluePrint)
    api.register_blueprint(AlphabetBluePrint)

    return app


    

